Apply second round of PR review fixes: NULL init, prev fix, unidirectional shape, format specifiers, doc typos, ISA guard, vscode path
Agent-Logs-Url: https://github.com/tkumata/prone-model/sessions/17087647-58b7-4e7a-93b2-f94bbb453c48

+++ components/espressif__esp-dl/audio/common/dl_audio_wav.cpp
@@ -1,5 +1,6 @@
 
 #include "dl_audio_wav.hpp"
+#include <inttypes.h>
 namespace dl {
 namespace audio {
 
@@ -124,8 +125,8 @@ void print_audio_info(dl_audio_t *data)
     printf("  Sample Rate: %d Hz\n", data->sample_rate);
     printf("  Channels: %d\n", data->channels);
     printf("  Bits per Sample: %d\n", data->bits_per_sample);
-    printf("  Length: %ld samples\n", data->length);
-    printf("  Data Size: %ld bytes\n", sizeof(int16_t) * data->length * data->channels);
+    printf("  Length: %" PRIu32 " samples\n", data->length);
+    printf("  Data Size: %zu bytes\n", sizeof(int16_t) * (size_t)data->length * data->channels);
 }
 
 } // namespace audio
